Rename ResultSetMetaDataIO -> ResultSetMetaDataSyntax
