+++ module/ldbc-dsl/src/main/scala/ldbc/dsl/internal/package.scala
@@ -4,5 +4,5 @@
 
 package ldbc.dsl
 
-trait internalSyntax extends internal.ConnectionSyntax
+trait internalSyntax extends internal.ConnectionSyntax, internal.ResultSetMetaDataSyntax
 package object internal extends internalSyntax
